refactor/break-out api request state
